+++ events/copyToClipboard.js
@@ -1,0 +1,5 @@
+const copyToClipboard = () => {
+  console.log("copyToClipboard running...");
+  const input = document.querySelector("#sharedURL");
+  navigator.clipboard.writeText(input.value);
+};