+++ prisma/migrations/20250314003415_created_shared_folders_model/migration.sql
@@ -1,0 +1,9 @@
+-- CreateTable
+CREATE TABLE "SharedFolders" (
+    "id" TEXT NOT NULL,
+    "createdAt" TIMESTAMP(3) NOT NULL DEFAULT CURRENT_TIMESTAMP,
+    "expiresAt" TIMESTAMP(3) NOT NULL,
+    "folderId" TEXT NOT NULL,
+
+    CONSTRAINT "SharedFolders_pkey" PRIMARY KEY ("id")
+);